Frontend - refacto WIP by creating a component by form step
